Split test files and move to their own folder
